CrashLog() support windows
