+++ log.go
@@ -7,7 +7,6 @@ import (
 	"log"
 	"os"
 	"sync"
-	"syscall"
 	"time"
 )
 
@@ -54,15 +53,6 @@ func init() {
 
 func Logger() *log.Logger {
 	return _log._log
-}
-
-func CrashLog(file string) {
-	f, err := os.OpenFile(file, os.O_WRONLY|os.O_CREATE|os.O_APPEND, 0666)
-	if err != nil {
-		log.Println(err.Error())
-	} else {
-		syscall.Dup2(int(f.Fd()), 2)
-	}
 }
 
 func SetLevel(level LogLevel) {
@@ -210,14 +200,14 @@ func (l *logger) rotate() error {
 }
 
 func (l *logger) doRotate(suffix string) error {
+	// Notice: Not check error, is this ok?
+	l.fd.Close()
+
 	lastFileName := l.fileName + "." + l.logSuffix
 	err := os.Rename(l.fileName, lastFileName)
 	if err != nil {
 		return err
 	}
-
-	// Notice: Not check error, is this ok?
-	l.fd.Close()
 
 	err = l.SetOutputByName(l.fileName)
 	if err != nil {
